% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssql.R
\name{mssql}
\alias{mssql}
\title{Get response from Microsoft SQL Server}
\usage{
mssql(con,
  query = "SELECT 200 AS Response, CURRENT_TIMESTAMP AS Relevance")
}
\arguments{
\item{con}{connection object of class \code{\link{mssql_connection}}
of \emph{sqsh} command.}

\item{query}{valid \emph{SQL}-expression as character}
}
\value{
\emph{Microsoft SQL server} response as \emph{data.table}
}
\description{
Send \emph{SQL}-query and get server response as \emph{data.table}
}
\details{
Connection problems are mostly processed and diagnostics is returned as
 \emph{data.table}. Error codes are \emph{HTTP} status codes.
}
\examples{
\dontrun{
# Create connection object:
conn <- mssql_connection(server = "10.100.104.142", domain = "SUEK",
 user = "user", password = "password")

# Ask for data:
 mssql(conn, "SELECT 200 AS V1, 300 AS V2")

# Output:
 #     V1  V2
 # 1: 200 300
}
}
\seealso{
\code{\link{mssql_connection}} for creating connection object

Other Database: \code{\link{mssql_connection}}
}
\concept{Database}
