% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{mssql_connection}
\alias{mssql_connection}
\title{Connect to a Microsoft SQL Server}
\usage{
mssql_connection(server = "", domain = "", user = "",
  password = "", tds_version = "8.0")
}
\arguments{
\item{server}{client hostname or IP-address. The value is associated with \emph{-H} option
of \emph{sqsh} command.}

\item{domain}{user domain name (if any)}

\item{user}{database username to connect to server. The value is associated with \emph{-U}
option of \emph{sqsh} command.}

\item{password}{password for username required to connect to server. The value is associated
with \emph{-P} option of \emph{sqsh} command.}

\item{tds_version}{Set the TDS version to use. Valid versions are \emph{4.0}, \emph{4.2},
\emph{4.6}, \emph{4.9.5}, \emph{5.0}, \emph{7.0} and \emph{8.0}.
The value is associated with \emph{-G} option of \emph{sqsh} command.}
}
\value{
Connection object holding requisites of connection
}
\description{
Connections to \emph{Microsoft SQL server} are executed on top of
 \href{https://www.freetds.org/}{FreeTDS} driver using \emph{sqsh} command line interface
 internally. Connections to \emph{Microsoft SQL server} are made each time
 the user sends the \emph{SQL}-request. The object created by the function
 holds requisites only. So you do not need to disconnect.

 Note that \emph{sqsh} client must be installed on the client's system.
}
\details{
Do not use domain name in \emph{user}, put it to \emph{domain}.
}
\examples{
\dontrun{
# Create connection object:
conn <- mssql_connection(server = "10.100.104.142", domain = "SUEK",
 user = "user", password = "password")

# Ask for data:
 mssql(conn, "SELECT 200 AS V1, 300 AS V2")

# Output:
 #     V1  V2
 # 1: 200 300
}
}
\seealso{
\code{\link{mssql}} for sending \emph{SQL-requests}

Other Database: \code{\link{mssql}}
}
\concept{Database}
